% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEO_to_Cloupe.R
\name{GEO_to_Cloupe}
\alias{GEO_to_Cloupe}
\title{GEO scRNA Downloader and Cloupe Converter}
\usage{
GEO_to_Cloupe(GEO_ID_List, File_Format)
}
\arguments{
\item{GEO_ID_List}{A list of strings that contain GEO accession codes}

\item{File_Format}{A string of the file format ending type}
}
\value{
Returns a merged Seurat object of all the different scRNA datasets downloaded, while also exporting cloupe files into the directory "CloupeFiles"
}
\description{
Automatically downloads scRNA files from GEO bases and converts them into cloupe file format
}
\examples{
merged_Seurat <- GEO_to_Cloupe(c("GSE108677"), ".h5");
merged_Suerat <- (c("GSE108677", "GSE128033"), "mtx.gz" );
}
