% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEOtoCloupe.R
\name{GEO_to_Cloupe}
\alias{GEO_to_Cloupe}
\title{GEO scRNA Downloader and Cloupe Converter}
\usage{
GEO_to_Cloupe(
  GEO_ID_List,
  Downloaded = FALSE,
  Integrate = FALSE,
  Resolution = 0.5,
  Merge = TRUE,
  Mitochondria = 20
)
}
\arguments{
\item{GEO_ID_List}{A list of strings that contain GEO accession codes}

\item{Downloaded}{A TRUE or FALSE that dictates whether the files have been downloaded prior. Useful for debugging and testing codes without having to re-download files.}

\item{Integrate}{A TRUE or FALSE that dictates whether to integrate the different scRNA datasets together}

\item{Resolution}{Set the value for UMAP clustering sensitivity from a value of 0.0 to 1.0 from less to more clusters}

\item{Merge}{A TRUE or FALSE that dictates whether you want to merge the multiple scRNA data}

\item{Mitochondria}{Set threshold for mitochondria QC metrics. Values set between 0 to 100.}
}
\value{
Returns a merged Seurat object of all the different scRNA datasets downloaded, while also exporting cloupe files into the directory "CloupeFiles"
}
\description{
Automatically downloads scRNA files from GEO bases and converts them into cloupe file format
}
\examples{
merged_Seurat <- GEO_to_Cloupe(c("GSE213338"));
merged_Suerat <- GEO_to_Cloupe(c("GSE213338", "GSE162733"), Downloaded = TRUE, Merge = FALSE);
}
